bug hunting
I am chasing a weird bug that generates  a hard fault after a while and stops the teensy. I found that it tends to occur if line 228 in controller.cpp is there. I hunted for a possible use of the heap in all the libraries but could not find one.